Format the first container on the landing page
The format is using flexbox to put an image place holder
and another container of text and a button side by side.

Eventually I need to make the image have an absolute size.
I don't want the image to stretch. It should stay the same size
but move position on the webpage when resizing the site.